+++ pacman-core/src/main/java/de/amr/games/pacman/controller/mspacman/MsPacManIntro.java
@@ -74,7 +74,7 @@ public class MsPacManIntro extends Fsm<IntroState, IntroData> {
 		public final Vector2i titlePosition = v2i(t(10), t(8));
 		public final Pulse blinking = new Pulse(30, true);
 		public final TickTimer lightsTimer = new TickTimer("lights-timer");
-		public final double actorSpeed = 1.1f;
+		public final float actorSpeed = 1.1f;
 		public int ghostIndex = 0;
 		public boolean creditVisible = false;
