+++ pacman-core/src/main/java/de/amr/games/pacman/model/common/actors/Entity.java
@@ -29,7 +29,7 @@ import static de.amr.games.pacman.model.common.world.World.tileAt;
 
 import java.util.Objects;
 
-import de.amr.games.pacman.lib.math.Vector2d;
+import de.amr.games.pacman.lib.math.Vector2f;
 import de.amr.games.pacman.lib.math.Vector2i;
 
 /**
@@ -40,15 +40,15 @@ import de.amr.games.pacman.lib.math.Vector2i;
 public class Entity {
 
 	protected boolean visible;
-	protected Vector2d position;
-	protected Vector2d velocity;
-	protected Vector2d acceleration;
+	protected Vector2f position;
+	protected Vector2f velocity;
+	protected Vector2f acceleration;
 
 	public Entity() {
 		visible = false;
-		position = Vector2d.ZERO;
-		velocity = Vector2d.ZERO;
-		acceleration = Vector2d.ZERO;
+		position = Vector2f.ZERO;
+		velocity = Vector2f.ZERO;
+		acceleration = Vector2f.ZERO;
 	}
 
 	public boolean isVisible() {
@@ -67,40 +67,40 @@ public class Entity {
 		visible = false;
 	}
 
-	public Vector2d position() {
+	public Vector2f position() {
 		return position;
 	}
 
-	public void setPosition(double x, double y) {
-		position = new Vector2d(x, y);
+	public void setPosition(float x, float y) {
+		position = new Vector2f(x, y);
 	}
 
-	public void setPosition(Vector2d position) {
+	public void setPosition(Vector2f position) {
 		this.position = Objects.requireNonNull(position, "Position of entity must not be null");
 	}
 
-	public Vector2d velocity() {
+	public Vector2f velocity() {
 		return velocity;
 	}
 
-	public void setVelocity(Vector2d velocity) {
+	public void setVelocity(Vector2f velocity) {
 		this.velocity = Objects.requireNonNull(velocity, "Velocity of entity must not be null");
 	}
 
-	public void setVelocity(double vx, double vy) {
-		velocity = new Vector2d(vx, vy);
+	public void setVelocity(float vx, float vy) {
+		velocity = new Vector2f(vx, vy);
 	}
 
-	public Vector2d acceleration() {
+	public Vector2f acceleration() {
 		return acceleration;
 	}
 
-	public void setAcceleration(Vector2d acceleration) {
+	public void setAcceleration(Vector2f acceleration) {
 		this.acceleration = Objects.requireNonNull(acceleration, "Acceleration of entity must not be null");
 	}
 
-	public void setAcceleration(double ax, double ay) {
-		acceleration = new Vector2d(ax, ay);
+	public void setAcceleration(float ax, float ay) {
+		acceleration = new Vector2f(ax, ay);
 	}
 
 	/**
@@ -122,12 +122,12 @@ public class Entity {
 	}
 
 	/** Center of "bounding box" (position stores upper left corner of bounding box). */
-	public Vector2d center() {
+	public Vector2f center() {
 		return position.plus(HTS, HTS);
 	}
 
 	/** Offset: (0, 0) if centered, range: [-4, +4) */
-	public Vector2d offset() {
+	public Vector2f offset() {
 		return position.minus(originOfTile(tile()));
 	}
 }
